move files to lib, update imports
