+++ src/lib/StockDataClient.ts
